Structure: Restructure lists
