+++ src/components/general/lists/List/List.component.tsx
@@ -3,22 +3,23 @@ import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
 
 interface ListProps<T> {
   /**
-   * Data, that shall be rendered in the list.
+   * Data, that shall be rendered in the List.
    */
   data: T;
+  /**
+   * Custom style for this List-container.
+   */
+  style?: StyleProp<ViewStyle>;
+
   /**
    * Renderer for the data. It excepts one argument of type T (the type of the data) and returns a ReactNode.
    * @param data data to be rendered
    */
   render?(data: T): React.ReactNode;
-  /**
-   * Custom style for this list-container.
-   */
-  style?: StyleProp<ViewStyle>;
 }
 
 /**
- * A simple display utility for displaying a list.
+ * A simple display utility for displaying a List.
  */
 export function List<T>({ style, render, data }: ListProps<T>) {
   // Use a very simple render function if none is provided
