+++ src/components/general/lists/List/index.ts
@@ -1,0 +1,1 @@
+export * from './List.component';
