+++ src/components/general/lists/SectionHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './SectionHeader.component';
