+++ src/screens/addPayments/searchList/searchList.screen.tsx
@@ -1,6 +1,6 @@
 import { PescaButton } from '@moneyboy/components/general/input/PescaButton';
 import { PescaInputField } from '@moneyboy/components/general/input/PescaInputField';
-import { ListItem } from '@moneyboy/components/general/lists/listItem';
+import { ListItem } from '@moneyboy/components/general/lists/ListItem';
 import { ScreenComponentProps } from '@moneyboy/components/general/navigation/pesca-navigator/pescaScreen';
 import { PescaContext } from '@moneyboy/contexts/pescaContext';
 import { useAuth } from '@moneyboy/hooks/useAuth';
