+++ src/screens/history/history.screen.tsx
@@ -1,4 +1,4 @@
-import { SectionHeader } from '@moneyboy/components/general/lists/sectionHeader';
+import { SectionHeader } from '@moneyboy/components/general/lists/SectionHeader';
 import { MoneyDiff, MoneyDiffProps } from '@moneyboy/components/general/payments/moneyDiff';
 import { Content } from '@moneyboy/components/general/structure/content';
 import { ViewBase } from '@moneyboy/components/general/structure/viewBase';
